moves API tests to Legacy->Two folder
