+++ tests/Http/Legacy/Two/SignupApiTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Http\Api;
+namespace Tests\Http\Api\Two;
 
 use Tests\TestCase;
 use DoSomething\Gateway\Blink;
